requirements.txt
python-dotenv==1.0.0
streamlit==1.31.1
langchain==0.1.11
langchain-openai==0.1.6
faiss-cpu==1.13.0
pypdf==4.0.1
pymupdf==1.23.8
sentence-transformers==2.2.2
scikit-learn==1.4.1
numpy==1.24.3
pyyaml==6.0
pytest==7.4.3
